+++ mobsf/StaticAnalyzer/models.py
@@ -96,7 +96,8 @@ class StaticAnalyzerIOS(models.Model):
     STRINGS = models.TextField(default=[])
     FIREBASE_URLS = models.TextField(default=[])
     APPSTORE_DETAILS = models.TextField(default={})
-    SECRETS = models.TextField(default={}) 
+    SECRETS = models.TextField(default=[])
+
 
 class StaticAnalyzerWindows(models.Model):
     FILE_NAME = models.CharField(max_length=260, default='')
