+++ mobsf/StaticAnalyzer/views/android/strings.py
@@ -5,35 +5,12 @@ import os
 
 from androguard.core.bytecodes import apk
 
-from mobsf.StaticAnalyzer.views.shared_func import url_n_email_extract
+from mobsf.StaticAnalyzer.views.shared_func import (
+    is_secret,
+    url_n_email_extract,
+)
 
 logger = logging.getLogger(__name__)
-
-
-def is_secret(inp):
-    inp = inp.lower()
-    """Check if captures string is a possible secret."""
-    iden = (
-        'api"', 'key"', 'api_', 'key_', 'secret"',
-        'password"', 'aws', 'gcp', 's3_', '_s3', 'secret_',
-        'token"', 'username"', 'user_name"', 'user"',
-        'bearer', 'jwt', 'certificate"', 'credential',
-        'azure', 'webhook', 'twilio_', 'bitcoin',
-        '_auth', 'firebase', 'oauth', 'authorization',
-        'private', 'pwd', 'session', 'token_',
-    )
-    not_string = (
-        'label_', 'text', 'hint', 'msg_', 'create_',
-        'message', 'new', 'confirm', 'activity_',
-        'forgot', 'dashboard_', 'current_', 'signup',
-        'sign_in', 'signin', 'title_', 'welcome_',
-        'change_', 'this_', 'the_', 'placeholder',
-        'invalid_', 'btn_', 'action_', 'prompt_',
-        'lable', 'hide_', 'old', 'update', 'error',
-        'empty', 'txt_', 'lbl_',
-    )
-    not_str = any(i in inp for i in not_string)
-    return any(i in inp for i in iden) and not not_str
 
 
 def strings_from_apk(app_file, app_dir, elf_strings):
