+++ mobsf/StaticAnalyzer/views/shared_func.py
@@ -396,3 +396,29 @@ def find_java_source_folder(base_folder: Path):
                             (base_folder / 'src',
                              'java', '*.java')]
                 if p[0].exists())
+
+
+def is_secret(inp):
+    inp = inp.lower()
+    """Check if captures string is a possible secret."""
+    iden = (
+        'api"', 'key"', 'api_', 'key_', 'secret"',
+        'password"', 'aws', 'gcp', 's3_', '_s3', 'secret_',
+        'token"', 'username"', 'user_name"', 'user"',
+        'bearer', 'jwt', 'certificate"', 'credential',
+        'azure', 'webhook', 'twilio_', 'bitcoin',
+        '_auth', 'firebase', 'oauth', 'authorization',
+        'private', 'pwd', 'session', 'token_',
+    )
+    not_string = (
+        'label_', 'text', 'hint', 'msg_', 'create_',
+        'message', 'new', 'confirm', 'activity_',
+        'forgot', 'dashboard_', 'current_', 'signup',
+        'sign_in', 'signin', 'title_', 'welcome_',
+        'change_', 'this_', 'the_', 'placeholder',
+        'invalid_', 'btn_', 'action_', 'prompt_',
+        'lable', 'hide_', 'old', 'update', 'error',
+        'empty', 'txt_', 'lbl_',
+    )
+    not_str = any(i in inp for i in not_string)
+    return any(i in inp for i in iden) and not not_str
