+++ mobsf/templates/pdf/ios_report.html
@@ -632,7 +632,26 @@
 										</table>
       {% endif %}
   
-     
+    {% if secrets %}
+      <h2><i class="fas fa-key"></i> HARDCODED SECRETS</h2>
+          <table class="basic">
+                  <thead>
+                  <tr>
+                    <th>POSSIBLE SECRETS</th>
+                  </tr>
+                </thead>
+                  <tbody>
+                  {% for sec in secrets %}
+                    <tr>
+                      <td style="word-break:break-all;">
+                      {{ sec }}
+                        </td>
+                    </tr>
+                  {% endfor %} 
+                    </tbody>
+                    </table>
+      {% endif %}
+
     {% if not appstore_details.error %}
     <h2><i class="fab fa-app-store-ios"></i> APP STORE INFORMATION </h2>
